Delegated factor analysis and factor ops to new objects
Delegated factor analysis and factor operations to
new objects. This new api model permits us to
abstract away operations from data required by it.
The new desined lacks principle for now it needs to
be sketched in a design document. There is already
a design.md document which contains the overall
description.
A diagram would be nice to have.