Heavy refactoring in abstract objects
Abstract objects now type check some of their
abstractmethods. Basically they call some of the
abstract methods in check_types function which
checks each method result for its type.
The check_types method then is called in
the constructor of the abstract class.
That way it is checked for all subclassing
objects.

The same idea should be implemented for abstractfactor
and abstractpgm objects.