Fixed linter errors and advanced in path
Fixed some of the linter errors. Now most of the errors are
about test setup being too complex, and some unused variables
in unfinished functions here and there.

Advanced also in path object. Now we check whether the edge list
satisfies conditions for being a path. Basically we check if
it has more than 2 heads or 2 tails. How to organize those
heads and tails depend on the edge type of the graph thus should
be left to more specific implementations. Mixed paths should also
be considered.