Factor refactoring more or less done
Factor.py contains reasonable methods now.
Not sure truncating it further would resolve anything.
The todo list in the log shows that the next area
of refactoring in line is random variables.
They need to be adapted to trait like architecture that
we are experimenting. The documentation also needs to
be updated radically as it has nothing to do with the
actual state of the api.