New random variable design
New random variable design is underway.
We loosen the difference between categorical random variable
and a numeric categorical random variable. We assume in most
categoricalops that random variables are numeric categorical
random variables, that is the output associated to the outcome
of an event is modeled as a number. NumCatRandVar methods
in categorical must transfer to categoricalops.py.
Once that is done, the new design will enter into code base
and the tests of old randomvariable.py would be distributed
accordingly.