Tests added continuing categoricalops tests
New tests are added to test_categoricalops.py. In order to track
methods that have an associated test, I added a test notice to
all functions of categoricalops. There still functions that
do not have a test.
The other thing is it might be interesting to divide the current
namespaces into modules. For example:
instead of NumCatRandomVarNumericOps it might be more interesting
to have
from categoricalops.numeric import NumericOps which contains
the same operations.