Assertions are the key to interpretation
Both Definetti and consequently Biagini uses a particularly unique
interpretation of the notion 'assertion'.
The assertion means that proposition following the assertion symbol
evaluates to true.
This means that event representing the proposition happenning
has only 1 as its outcome.
It is tricky to represent this, because possible events tend to
yield 0 at least in some operations concerning exhaustivity.
If |- E1 + E2 + E3 + ... >= 1 is the assertion that E1 + E2 + E3
are exhaustive, this means that E1 + E2 + E3 >= 1 happens.
If none of the events are certain, there will be a 0 in the
arithmetic sum of the events.
Another tricky bit is the complements. If |- X + ~X = 1 is an
assertion, this should hold for even possible cases, since when
X=0, the ~X=1 and vice versa. But once we consider the ~X as an
independent random number, we again fall to the trap of
having 0s in the arithmetic sum of events.