Kolmogorov vs DeFinetti
The measure theoretic probability interpretation vs DeFinetti's
subjective interpretation. It is quite difficult to choose.
The discrete random number follows DeFinetti's calculus for
random numbers, while the Kolmogorov's measure theory based
interpretation unifies the discrete and continuous random numbers.
I think, I'll stick to measure theoretic formulation.
This means the possible outcome set would need to be a member of
sigma field and since we deal with numeric outcomes in general
they would need to be a borel set.
There is a check that is implemented for sigma fields in
valuefunc/setops.py but there is nothing for checking borel sets.

Here are some useful references:
- Shao, 2010: easier to follow, the pace is quite well
- Kallenberg, 2002: too many stuff introduced too fast, but nice
for p-system, l-system definitions.
- LeGall, 2022: not bad for cross reading with Shao, but sometimes
supposes too much familiarity with domain and gloss over certain
terms.
- Brémaud, 2017: has a nice blend of strictness and intuition.
It can be used to bridge current code to measure theoretic
framework.
- Mendelhall, 2020: is good for finding examples and for unittests