Here is the plan
The roadmap for sample space and measure set implementation is
clear now.
We need to structures: Intervals and Countables.
Countables are more or less understandable, they are enumerable
sets, and can be implemented as frozenset's normal sets etc.
It is also easier to implement set operations on them.
Intervals are more delicate. They are not enumerable but bounded
in most cases. Implementing set operations require more
consideration, like whether the interval is open or not. The
R line can be implemented as an interval if can overload the
operations of NumericValue for float("inf") and float("-inf").
The required algebra is defined in:
Shao 2010, p. 3.

Once we can define the real line as an interval we can apply
setops in valuefunc to both real line and to enumerable sets.
This also requires us to implement set operations used in setops
to be implemented by the interval class.