Interval setops
Interval set operations are written, we can now add tests for it
and change setops in valuefunc to consider intervals and/or
sets of intervals. We should probably define subset values as a
degenerate interval, or at least permit an exchange between the
two using this degenerate interval notion. Then, it is all about
testing these functions